+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **VorbisComments**:
   - Keys will now be verified according to spec before insertion
   - Getters will now case-insensitively search for keys
+  - `TRACKNUM` will now be considered in the `Accessor::*track` methods
 
 ### Fixed
 - **ID3v2**: